pyproject.toml
[tool.poetry]
name = "voice-in"
version = "0.2.0"
description = "AI-powered Voice Dictation Tool"
authors = ["Your Name <your@email.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
PyQt6 = "^6.7.0"
sounddevice = "^0.5.0"
numpy = "^1.26.0"
scipy = "^1.13.0"
groq = "^0.9.0"
google-generativeai = "^0.7.0"
pynput = "^1.7.0"
python-dotenv = "^1.0.0"
faster-whisper = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
